stop rendering some class docs in wrong place (#8095)
* stop rendering some class docs in wrong place

Docs generated for providers.yandex.hooks incorrectly include docs
for airflow.exceptions.AirflowException and
airflow.hooks.base_hook.BaseHook in the yandex module, as if
those classes had been defined in that module.

This is almost certainly a bug in autoapi or one of the libraries
it uses, but I haven't tracked down the root cause. In the meantime,
importing the modules and then referring to the classes using
modulename.Classname avoids the issue.
