+++ tests/test_project_structure.py
@@ -46,7 +46,8 @@ MISSING_TEST_FILES = {
     'tests/providers/microsoft/mssql/hooks/test_mssql.py',
     'tests/providers/oracle/operators/test_oracle.py',
     'tests/providers/qubole/hooks/test_qubole.py',
-    'tests/providers/samba/hooks/test_samba.py'
+    'tests/providers/samba/hooks/test_samba.py',
+    'tests/providers/yandex/hooks/test_yandex.py'
 }
